webserver/requirements.txt
redis
python-multipart
Pillow
numpy